+++ packages/terra-application-layout/src/terra-dev-site/test/application-layout/ApplicationLayoutWithIcons.test.jsx
@@ -223,7 +223,7 @@ const userData = {
  */
 const nameConfig = Object.freeze({
   title: 'Test Application',
-  accessory: <Image variant="rounded" src="https://github.com/cerner/terra-framework/raw/master/terra.png" alt="App logo" height="26px" width="26px" />,
+  accessory: <Image variant="rounded" src="https://github.com/cerner/terra-framework/raw/main/terra.png" alt="App logo" height="26px" width="26px" />,
 });
 
 class ApplicationLayoutTest extends React.Component {