+++ packages/terra-brand-footer/tests/jest/BrandFooter.test.jsx
@@ -76,7 +76,7 @@ describe('BrandFooter', () => {
   it('should render with provided content left', () => {
     const wrapper = shallow(<BrandFooter
       contentLeft={
-        <img src="https://github.com/cerner/terra-core/raw/master/terra.png" alt="terra-ui" height="40px" width="40px" />
+        <img src="https://github.com/cerner/terra-core/raw/main/terra.png" alt="terra-ui" height="40px" width="40px" />
         }
     />);
     expect(wrapper).toMatchSnapshot();
@@ -117,7 +117,7 @@ describe('BrandFooter', () => {
           },
         ]}
         contentLeft={
-          <img src="https://github.com/cerner/terra-core/raw/master/terra.png" alt="terra-ui" height="40px" width="40px" />
+          <img src="https://github.com/cerner/terra-core/raw/main/terra.png" alt="terra-ui" height="40px" width="40px" />
         }
       />
     );
@@ -182,7 +182,7 @@ describe('BrandFooter', () => {
           },
         ]}
         contentLeft={
-          <img src="https://github.com/cerner/terra-core/raw/master/terra.png" alt="terra-ui" height="40px" width="40px" />
+          <img src="https://github.com/cerner/terra-core/raw/main/terra.png" alt="terra-ui" height="40px" width="40px" />
       }
         contentBottom={(
           <p>
@@ -213,7 +213,7 @@ describe('BrandFooter', () => {
           },
         ]}
         contentLeft={
-          <img src="https://github.com/cerner/terra-core/raw/master/terra.png" alt="terra-ui" height="40px" width="40px" />
+          <img src="https://github.com/cerner/terra-core/raw/main/terra.png" alt="terra-ui" height="40px" width="40px" />
       }
       />
     );
@@ -271,7 +271,7 @@ describe('BrandFooter', () => {
         },
       ]}
       contentLeft={
-        <img src="https://github.com/cerner/terra-core/raw/master/terra.png" alt="terra-ui" height="40px" width="40px" />
+        <img src="https://github.com/cerner/terra-core/raw/main/terra.png" alt="terra-ui" height="40px" width="40px" />
         }
       contentBottom={(
         <p>
@@ -296,7 +296,7 @@ describe('BrandFooter', () => {
         </p>
 )}
       contentLeft={
-        <img src="https://github.com/cerner/terra-core/raw/master/terra.png" alt="terra-ui" height="40px" width="40px" />
+        <img src="https://github.com/cerner/terra-core/raw/main/terra.png" alt="terra-ui" height="40px" width="40px" />
         }
     />);
     expect(wrapper).toMatchSnapshot();
@@ -322,7 +322,7 @@ describe('BrandFooter', () => {
   it('should render with content bottom, content right and content left', () => {
     const wrapper = shallow(<BrandFooter
       contentLeft={
-        <img src="https://github.com/cerner/terra-core/raw/master/terra.png" alt="terra-ui" height="40px" width="40px" />
+        <img src="https://github.com/cerner/terra-core/raw/main/terra.png" alt="terra-ui" height="40px" width="40px" />
         }
       contentBottom={(
         <p>
@@ -342,7 +342,7 @@ describe('BrandFooter', () => {
   it('should render with content right and content left', () => {
     const wrapper = shallow(<BrandFooter
       contentLeft={
-        <img src="https://github.com/cerner/terra-core/raw/master/terra.png" alt="terra-ui" height="40px" width="40px" />
+        <img src="https://github.com/cerner/terra-core/raw/main/terra.png" alt="terra-ui" height="40px" width="40px" />
         }
       contentRight={(
         <div>