+++ packages/terra-time-input/src/TimeInput.jsx
@@ -21,12 +21,10 @@ const propTypes = {
   /**
    * Custom input attributes that apply to the hour, minute, and second inputs.
    */
-  // eslint-disable-next-line react/forbid-prop-types
   inputAttributes: PropTypes.object,
   /**
    * Custom input attributes to apply to the hour input
    */
-  // eslint-disable-next-line react/forbid-prop-types
   hourAttributes: PropTypes.object,
   /**
   * @private
@@ -48,7 +46,6 @@ const propTypes = {
   /**
    * Custom input attributes to apply to the minutes input
    */
-  // eslint-disable-next-line react/forbid-prop-types
   minuteAttributes: PropTypes.object,
   /**
    * Name of the time input. The name should be unique.
@@ -79,7 +76,6 @@ const propTypes = {
   /**
    * Custom input attributes to apply to the seconds input
    */
-  // eslint-disable-next-line react/forbid-prop-types
   secondAttributes: PropTypes.object,
   /**
    * Whether the input for seconds should be displayed or not. If true then the second field must have a valid
@@ -175,7 +171,6 @@ class TimeInput extends React.Component {
       if (hour) {
         const parsedHour = TimeUtil.parseTwelveHourTime(hour, this.anteMeridiem, this.postMeridiem);
         hour = parsedHour.hourString;
-        // eslint-disable-next-line prefer-destructuring
         meridiem = parsedHour.meridiem;
       } else {
         meridiem = this.anteMeridiem;
@@ -214,7 +209,6 @@ class TimeInput extends React.Component {
       if (hour) {
         const parsedHour = TimeUtil.parseTwelveHourTime(hour, this.anteMeridiem, this.postMeridiem);
         hour = parsedHour.hourString;
-        // eslint-disable-next-line prefer-destructuring
         meridiem = parsedHour.meridiem;
       }
     }
@@ -730,7 +724,6 @@ class TimeInput extends React.Component {
       { 'initial-focus': this.state.secondInitialFocused },
     ]);
 
-    /* eslint-disable jsx-a11y/no-static-element-interactions */
     return (
       <div
         {...customProps}