+++ packages/terra-application-utility/CHANGELOG.md
@@ -3,6 +3,9 @@ Changelog
 
 Unreleased
 ----------
+
+2.36.0 - (July 14, 2020)
+------------------
 ### Changed
 * updated `master` references to `main`.
 * Remove unused disable eslint directives